rename dataset to data, restructure data section
