+++ data/home.yml
@@ -10,14 +10,14 @@ features:
     description: |
       We aggregate and enrich open access research papers from around the world
 
-      [Read about our data](~dataset)
+      [Read about our data](~data)
     picture: /static/images/globe.svg
   - title: Unique APIs
     description: |
       We provide seamless access to content and data, through our
       unique [APIs](~services/api)
 
-      **Perfect for [text mining](~dataset)!**
+      **Perfect for [text mining](~data)!**
     picture: /static/images/api.svg
   - title: Powerful Services
     description: |
