--- pages/services/dataset.jsx
@@ -1,5 +1,0 @@
-import { ServicePage } from 'components/pages'
-import page from 'data/services/dataset.yml'
-import freePackage from 'data/services/free-package.yml'
-
-export default ServicePage.create(page, freePackage)
